requirements.txt
streamlit
openai
tenacity
termcolor
pywin32; platform_system=="Windows"
python-docx
gspread
oauth2client

